Atualizado códigos para que endpoints funcionem corretamente

+++ norktown_carford/app/controllers/owner_controller.py
@@ -1,3 +1,4 @@
+from flask import jsonify
 from ..models.owner import Owner
 from .. import db
 
@@ -21,7 +22,10 @@ def update_owner(owner_id, data):
 def delete_owner(owner_id):
     owner = Owner.query.get(owner_id)
     if owner:
+        # Verificar se existem carros associados
+        if owner.cars:
+            return jsonify({'error': "Não foi possível deletar owner pois ainda há carros no seu nome"}), 400
         db.session.delete(owner)
         db.session.commit()
-        return True
-    return False
+        return jsonify({'message': 'Owner deletado com sucesso'}), 200
+    return jsonify({'error': 'Owner não encontrado'}), 404
